Add files via upload
Uploaded a ZIP dataset for my systematic review titled "Health-seeking behaviours of immigrants, asylum seekers, and refugees in Europe: a systematic review of peer-reviewed articles". This dataset includes:

- CSV file containing detailed information on all articles reviewed in the initial screening phase.
- Abstracts of each article for quick reference.
- Decisions on whether articles were included or excluded.
- Labels and exclusion reasons for each article, providing clarity on the screening process.
- User notes offering insights into the review methodology and considerations.

This dataset is pivotal for understanding the scope and depth of the initial literature survey and the decision-making criteria applied in the systematic review process.